# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Univibe API",
        "description": "API documentation for Univibe project",
        "contact": {
            "name": "Your Name",
            "email": "your.email@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/about-us/titles": {
            "post": {
                "tags": [
                    "About Us"
                ],
                "summary": "Create an About Us title",
                "description": "Creates a new About Us title and returns the created title",
                "operationId": "ea2c8e067ebd71033b7c41de081b7a47",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "About Our Company"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "About Us title created successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/about-us/{aboutUsId}/details": {
            "put": {
                "tags": [
                    "About Us"
                ],
                "summary": "Update an About Us detail",
                "description": "Updates the specified detail for the About Us entry",
                "operationId": "7b6cde587d861d97a15a5aa7f6a73a37",
                "parameters": [
                    {
                        "name": "aboutUsId",
                        "in": "path",
                        "description": "ID of the About Us entry",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description for About Us"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "About Us detail updated successfully"
                    },
                    "404": {
                        "description": "About Us entry not found"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "About Us"
                ],
                "summary": "Add detail to About Us",
                "description": "Creates a new detail for the specified About Us entry",
                "operationId": "9efd100d1c588c056741a4faf592a09f",
                "parameters": [
                    {
                        "name": "aboutUsId",
                        "in": "path",
                        "description": "ID of the About Us entry",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "We are dedicated to excellence..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "About Us detail added successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "About Us entry not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/about-us": {
            "get": {
                "tags": [
                    "About Us"
                ],
                "summary": "Get all About Us entries with details",
                "description": "Retrieves all About Us entries along with their details",
                "operationId": "5ef71d894f1df7290a385eb3d33fa0af",
                "responses": {
                    "200": {
                        "description": "About Us entries retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/about-us/{aboutUsId}": {
            "get": {
                "tags": [
                    "About Us"
                ],
                "summary": "Get a single About Us entry with details",
                "description": "Retrieves a single About Us entry by ID along with its details",
                "operationId": "6e0f3299f75e8c94c70735ae866edc37",
                "parameters": [
                    {
                        "name": "aboutUsId",
                        "in": "path",
                        "description": "ID of the About Us entry",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "About Us entry retrieved successfully"
                    },
                    "404": {
                        "description": "About Us entry not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/about-us/{id}": {
            "put": {
                "tags": [
                    "About Us"
                ],
                "summary": "Update an About Us title",
                "description": "Updates the specified About Us title",
                "operationId": "08b2af0c343f588c201f7c1460647bd1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the About Us entry",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated About Us Title"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "About Us updated successfully"
                    },
                    "404": {
                        "description": "About Us entry not found"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
...
}